PDF/UA-2. Add new methods. Remove unnecessary methods
