+++ pdfbox-validation-model/src/test/java/org/verapdf/model/impl/pb/cos/PBCosTrailerTest.java
@@ -61,13 +61,6 @@ public class PBCosTrailerTest extends BaseTest {
     }
 
     @Test
-    public void testCatalogLink() {
-        List<? extends Object> catalog = actual.getLinkedObjects(PBCosTrailer.CATALOG);
-        Assert.assertEquals(catalog.size(), 1);
-        Assert.assertTrue(catalog.get(0) instanceof CosIndirect);
-    }
-
-    @Test
     public void testEncryptLink() {
         List<? extends Object> encrypt = actual.getLinkedObjects(PBCosTrailer.ENCRYPT);
         Assert.assertEquals(encrypt.size(), 1);
